proj restructure
