+++ cmd/run.go
@@ -9,10 +9,10 @@ import (
 	"os"
 	"strings"
 
-	"github.com/gregriff/ducky/models"
-	"github.com/gregriff/ducky/models/anthropic"
-	"github.com/gregriff/ducky/models/openai"
-	"github.com/gregriff/ducky/tui"
+	tui "github.com/gregriff/ducky/internal"
+	"github.com/gregriff/ducky/internal/models"
+	"github.com/gregriff/ducky/internal/models/anthropic"
+	"github.com/gregriff/ducky/internal/models/openai"
 	"github.com/spf13/cobra"
 	"github.com/spf13/viper"
 	"golang.org/x/term"
