+++ internal/chat/scrollback.go
