+++ internal/tui.go
@@ -10,12 +10,12 @@ import (
 	"github.com/charmbracelet/bubbles/v2/viewport"
 	tea "github.com/charmbracelet/bubbletea/v2"
 	"github.com/charmbracelet/lipgloss/v2"
-	"github.com/gregriff/ducky/models"
-	"github.com/gregriff/ducky/models/anthropic"
-	"github.com/gregriff/ducky/models/openai"
-	chat "github.com/gregriff/ducky/tui/chat"
-	styles "github.com/gregriff/ducky/tui/styles"
-	"github.com/gregriff/ducky/utils"
+	chat "github.com/gregriff/ducky/internal/chat"
+	"github.com/gregriff/ducky/internal/models"
+	"github.com/gregriff/ducky/internal/models/anthropic"
+	"github.com/gregriff/ducky/internal/models/openai"
+	styles "github.com/gregriff/ducky/internal/styles"
+	"github.com/gregriff/ducky/internal/utils"
 	zone "github.com/lrstanley/bubblezone/v2"
 	"github.com/muesli/reflow/wordwrap"
 )
