+++ internal/utils/math.go
